wexiv 1.1 Warning Massive Commit!!! Mostly trying to keep closer to upstream code. A few fixes. Iptc.cpp readData bug where I forgotten .key()

+++ src/sonymn_int.hpp
@@ -1,34 +1,11 @@
-// ***************************************************************** -*- C++ -*-
-/*
- * Copyright (C) 2004-2021 Exiv2 authors
- * This program is part of the Exiv2 distribution.
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, 5th Floor, Boston, MA 02110-1301 USA.
- */
+// SPDX-License-Identifier: GPL-2.0-or-later
+
 #ifndef SONYMN_INT_HPP_
 #define SONYMN_INT_HPP_
 
 // *****************************************************************************
 // included header files
-#include "tags.hpp"
 #include "tiffcomposite_int.hpp"
-#include "types.hpp"
-
-// + standard includes
-#include <math.h>
-#include <string>
 
 namespace Exiv2 {
 namespace Internal {
@@ -96,8 +73,8 @@ class SonyMakerNote {
 
 };  // class SonyMakerNote
 
-DataBuf sonyTagDecipher(uint16_t, const byte*, uint32_t, TiffComponent* const);
-DataBuf sonyTagEncipher(uint16_t, const byte*, uint32_t, TiffComponent* const);
+DataBuf sonyTagDecipher(uint16_t, const byte*, size_t, TiffComponent* const);
+DataBuf sonyTagEncipher(uint16_t, const byte*, size_t, TiffComponent* const);
 
 }  // namespace Internal
 }  // namespace Exiv2
